# src/main/webapp/public/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "AverageStats" : {
        "type" : "object",
        "properties" : {
          "pricePerUnit" : {
            "type" : "number"
          },
          "distance" : {
            "type" : "number"
          },
          "fuelCostPerKm" : {
            "type" : "number"
          },
          "fuelConsumption" : {
            "type" : "number"
          },
          "maintenanceCost" : {
            "type" : "number"
          },
          "costPerKm" : {
            "type" : "number"
          }
        }
      },
...
      "BillCategory" : {
        "type" : "string",
        "enum" : [ "FINANCE", "INSURANCE", "OTHER" ]
      },
      "BillInterval" : {
        "type" : "string",
        "enum" : [ "ONCE", "MONTHLY", "QUARTERLY", "SEMI_ANNUALLY", "ANNUALLY" ]
      },
...
      "BillPagedMiscellaneousBill" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MiscellaneousBill"
            }
          }
        }
      },
      "BillPagedRecurringBill" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "integer",
            "format" : "int64"
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RecurringBill"
            }
          }
        }
      },
...
      "DashboardStats" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "$ref" : "#/components/schemas/TotalStats"
          },
          "average" : {
            "$ref" : "#/components/schemas/AverageStats"
          },
          "hiLo" : {
            "$ref" : "#/components/schemas/HiLoStats"
          }
        }
      },
...
      "HiLo" : {
        "type" : "object",
        "properties" : {
          "min" : {
            "type" : "number"
          },
          "max" : {
            "type" : "number"
          },
          "scale" : {
            "type" : "integer",
            "format" : "int32",
            "writeOnly" : true
          }
        }
      },
      "HiLoStats" : {
        "type" : "object",
        "properties" : {
          "pricePerUnit" : {
            "$ref" : "#/components/schemas/HiLo"
          },
          "distance" : {
            "$ref" : "#/components/schemas/HiLo"
          },
          "fuelCostPerKm" : {
            "$ref" : "#/components/schemas/HiLo"
          },
          "consumption" : {
            "$ref" : "#/components/schemas/HiLo"
          },
          "fuelCost" : {
            "$ref" : "#/components/schemas/HiLo"
          },
          "maintenanceCost" : {
            "$ref" : "#/components/schemas/HiLo"
          }
        }
      },
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
...
      "MiscellaneousBill" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "year" : {
            "type" : "integer",
            "format" : "int32"
          },
          "description" : {
            "type" : "string"
          },
          "total" : {
            "type" : "number"
          }
        }
      },
      "MiscellaneousBillInput" : {
        "type" : "object",
        "properties" : {
          "total" : {
            "type" : "number"
          },
          "vatRate" : {
            "type" : "integer"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "description" : {
            "type" : "string"
          }
        }
      },
      "RecurringBill" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "nextDueDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "category" : {
            "$ref" : "#/components/schemas/BillCategory"
          },
          "startDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "endDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "amount" : {
            "type" : "number"
          },
          "total" : {
            "type" : "number"
          },
          "finished" : {
            "type" : "boolean"
          }
        }
      },
      "RecurringBillInput" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "billInterval" : {
            "$ref" : "#/components/schemas/BillInterval"
          },
          "category" : {
            "$ref" : "#/components/schemas/BillCategory"
          },
          "startDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "endDate" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "amount" : {
            "type" : "number"
          }
        }
      },
      "TotalStats" : {
        "type" : "object",
        "properties" : {
          "unit" : {
            "type" : "number"
          },
          "trackedDistance" : {
            "type" : "number"
          },
          "fuelTotal" : {
            "type" : "number"
          },
          "maintenanceTotal" : {
            "type" : "number"
          },
          "miscellaneousTotal" : {
            "type" : "number"
          },
          "total" : {
            "type" : "number"
          },
          "fuelBills" : {
            "type" : "integer",
            "format" : "int32"
          },
          "maintenanceBills" : {
            "type" : "integer",
            "format" : "int32"
          },
          "miscellaneousBills" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "type" : "openIdConnect",
        "description" : "Authentication"
      }
    }
  },
  "paths" : {
...
    "/api/bill/{carId}/miscellaneous" : {
      "parameters" : [ {
        "name" : "carId",
        "in" : "path",
        "required" : true,
        "schema" : {
          "type" : "integer",
          "format" : "int64"
        }
      } ],
      "put" : {
        "operationId" : "addNewMiscellaneousBill",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MiscellaneousBillInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "description" : "Bill created."
          },
          "409" : {
            "description" : "Car is not for your user."
          },
          "400" : {
            "description" : "Validation failed, please check the response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : { }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong while saving. Please ask the server admin for help."
          }
        },
        "summary" : "Add New Bill",
        "tags" : [ "Miscellaneous Resource" ]
      }
    },
    "/api/bill/{carId}/miscellaneous/all" : {
      "parameters" : [ {
        "name" : "carId",
        "in" : "path",
        "required" : true,
        "schema" : {
          "type" : "integer",
          "format" : "int64"
        }
      } ],
      "get" : {
        "description" : "Gets all bills for specified car, Pages starting at 1",
        "operationId" : "getAllMiscellaneousBills",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        }, {
          "name" : "year",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Bills found.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BillPagedMiscellaneousBill"
                }
              }
            }
          }
        },
        "summary" : "Get All My Bills",
        "tags" : [ "Miscellaneous Resource" ]
      }
    },
    "/api/bill/{carId}/miscellaneous/years" : {
      "parameters" : [ {
        "name" : "carId",
        "in" : "path",
        "required" : true,
        "schema" : {
          "type" : "integer",
          "format" : "int64"
        }
      } ],
      "get" : {
        "description" : "Gets all years of bills for specified car",
        "operationId" : "getAllMiscellaneousBillYears",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Bills found and years extracted.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "integer",
                    "format" : "int32"
                  }
                }
              }
            }
          }
        },
        "summary" : "Get All My Bills",
        "tags" : [ "Miscellaneous Resource" ]
      }
    },
    "/api/bill/{carId}/recurring" : {
      "get" : {
        "description" : "Gets all recurring bills for specified car",
        "operationId" : "getAllRecurringBills",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "onlyRunning",
          "in" : "query",
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 1
          }
        }, {
          "name" : "size",
          "in" : "query",
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BillPagedRecurringBill"
                }
              }
            }
          }
        },
        "summary" : "Get All Recurring Bills",
        "tags" : [ "Recurring Resource" ]
      },
      "put" : {
        "operationId" : "addNewRecurringBill",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecurringBillInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RecurringBill"
                }
              }
            }
          },
          "400" : {
            "description" : "Validation failed, please check the response",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : { }
                }
              }
            }
          }
        },
        "summary" : "Add New Recurring Bill",
        "tags" : [ "Recurring Resource" ]
      }
    },
    "/api/bill/{carId}/recurring/{billId}" : {
      "delete" : {
        "operationId" : "deleteRecurringBill",
        "parameters" : [ {
          "name" : "billId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "202" : {
            "description" : "Successful deleted"
          },
          "400" : {
            "description" : "Not able to delete"
          }
        },
        "summary" : "Delete Recurring Bill",
        "tags" : [ "Recurring Resource" ]
      }
    },
...
    "/api/stats/{carId}" : {
      "get" : {
        "operationId" : "getDashboardStats",
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "required" : false
        }, {
          "name" : "to",
          "in" : "query",
          "schema" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "required" : false
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DashboardStats"
                }
              }
            }
          }
        },
        "summary" : "Get Stats",
        "tags" : [ "Stats Resource" ]
      }
    },
...
  },
  "info" : {
    "title" : "car-ledger API",
    "version" : "0.1.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}